+++ packages/pc/tests/unit/graph/svg-spec.ts
@@ -1576,7 +1576,7 @@ describe('layouts', () => {
       graph.destroy();
     });
   });
-  it('change layout', () => {
+  it('change layout', (done) => {
     const graph = new Graph({
       container: div,
       width: 500,
@@ -1596,10 +1596,11 @@ describe('layouts', () => {
       type: 'force',
     });
 
-    Promise.resolve().then(() => {
+    setTimeout(() => {
       expect(graph.get('layoutController').layoutMethods[0].type).toBe('force');
       graph.destroy();
-    });
+      done();
+    }, 100);
   });
   it('subgraph layout', () => {
     const graph = new Graph({
@@ -2024,14 +2025,13 @@ describe('plugins', () => {
     setTimeout(() => {
       const minimapGroup = minimap2.get('canvas').get('children')[0];
       expect(minimapGroup.get('children').length).toBe(10);
-
       const viewport = minimap2.get('viewport');
-      expect(viewport.style.width).toBe('39.8406px');
-      expect(viewport.style.height).toBe('39.8406px');
-      expect(viewport.style.left).toBe('42.8287px');
-      expect(viewport.style.top).toBe('25.6574px');
+      expect(viewport.style.width).toBe('41.3907px');
+      expect(viewport.style.height).toBe('37.351px');
+      expect(viewport.style.left).toBe('58.6093px');
+      expect(viewport.style.top).toBe('42.649px');
       graph2.destroy();
-    }, 100);
+    }, 150);
   });
   it('minimap keyShape', () => {
     const minimap = new G6.Minimap({
@@ -2057,15 +2057,13 @@ describe('plugins', () => {
     setTimeout(() => {
       const minimapGroup = minimap.get('canvas').get('children')[0];
       expect(minimapGroup.get('children').length).toBe(10);
-
       const viewport = minimap.get('viewport');
-
-      expect(viewport.style.width).toBe('39.8406px');
-      expect(viewport.style.height).toBe('39.8406px');
-      expect(viewport.style.left).toBe('41.5787px');
-      expect(viewport.style.top).toBe('29.5618px');
+      expect(viewport.style.width).toBe('42.8115px');
+      expect(viewport.style.height).toBe('30.6977px');
+      expect(viewport.style.left).toBe('57.1885px');
+      expect(viewport.style.top).toBe('49.3023px');
       graph.destroy();
-    }, 100);
+    }, 150);
   });
 
   it('edge bundling', () => {